modify Reply-To header
Use the existing Reply-To header value, if it exists. This is in case the
sender has specified their own Reply-To value.

If a Reply-To header doesn't exist, use the From address. In some cases,
this address can be different than the senders address, which is what was
previously used.

One case I referred to above is a Google Apps account with an alias
domain. Mail sent from an alias domain will be from the alias, but sent by
the primary domain.